requirements.txt
Django==5.0
djangorestframework==3.15.1
PyJWT==2.8.0
bcrypt==4.1.2
drf-spectacular==0.27.2
python-dotenv==1.2.1
# psycopg2-binary
